Compile yml tests
