+++ docs/src/test/java/org/elasticsearch/smoketest/BaseDocsClientYamlTestSuiteCase.java
@@ -1,52 +1,26 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
 package org.elasticsearch.smoketest;
 
-import com.carrotsearch.randomizedtesting.annotations.Name;
-import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
-import com.carrotsearch.randomizedtesting.annotations.TimeoutSuite;
 import org.apache.http.HttpHost;
 import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.TimeUnits;
 import org.elasticsearch.Version;
 import org.elasticsearch.client.RestClient;
+import org.elasticsearch.common.CheckedFunction;
 import org.elasticsearch.common.ParseField;
 import org.elasticsearch.common.xcontent.ConstructingObjectParser;
-import org.elasticsearch.common.xcontent.NamedXContentRegistry;
 import org.elasticsearch.common.xcontent.XContentLocation;
 import org.elasticsearch.common.xcontent.XContentParser;
-import org.elasticsearch.common.xcontent.XContentParser.Token;
 import org.elasticsearch.test.rest.ESRestTestCase;
+import org.elasticsearch.test.rest.yaml.BaseCompiledYamlTestCase;
 import org.elasticsearch.test.rest.yaml.ClientYamlDocsTestClient;
-import org.elasticsearch.test.rest.yaml.ClientYamlTestCandidate;
 import org.elasticsearch.test.rest.yaml.ClientYamlTestClient;
 import org.elasticsearch.test.rest.yaml.ClientYamlTestExecutionContext;
 import org.elasticsearch.test.rest.yaml.ClientYamlTestResponse;
-import org.elasticsearch.test.rest.yaml.ESClientYamlSuiteTestCase;
 import org.elasticsearch.test.rest.yaml.restspec.ClientYamlSuiteRestSpec;
 import org.elasticsearch.test.rest.yaml.section.ExecutableSection;
 import org.junit.After;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -57,36 +31,18 @@ import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static org.elasticsearch.common.xcontent.ConstructingObjectParser.constructorArg;
 
-//The default 20 minutes timeout isn't always enough, please do not increase further than 30 before analyzing what makes this suite so slow
-@TimeoutSuite(millis = 30 * TimeUnits.MINUTE)
-public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
+public abstract class BaseDocsClientYamlTestSuiteCase extends BaseCompiledYamlTestCase {
 
-    private final ClientYamlTestCandidate testCandidate;
-
-    public DocsClientYamlTestSuiteIT(@Name("yaml") ClientYamlTestCandidate testCandidate) {
-        super(testCandidate);
-        this.testCandidate = testCandidate;
-    }
-
-    @ParametersFactory
-    public static Iterable<Object[]> parameters() throws Exception {
-//        List<NamedXContentRegistry.Entry> entries = new ArrayList<>(ExecutableSection.DEFAULT_EXECUTABLE_CONTEXTS.size() + 1);
-//        entries.addAll(ExecutableSection.DEFAULT_EXECUTABLE_CONTEXTS);
-//        entries.add(new NamedXContentRegistry.Entry(ExecutableSection.class,
-//                new ParseField("compare_analyzers"), CompareAnalyzers::parse));
-//        NamedXContentRegistry executableSectionRegistry = new NamedXContentRegistry(entries);
-//        return ESClientYamlSuiteTestCase.createParameters(executableSectionRegistry);
-        return Collections.emptyList();
+    protected BaseDocsClientYamlTestSuiteCase() throws IOException {
     }
 
     @Override
     protected void afterIfFailed(List<Throwable> errors) {
         super.afterIfFailed(errors);
-        String name = getTestName().split("=")[1];
-        name = name.substring(0, name.length() - 1);
+        String name = getTestName();
         name = name.replaceAll("/([^/]+)$", ".asciidoc:$1");
         logger.error("This failing test was generated by documentation starting at {}. It may include many snippets. "
-                + "See docs/README.asciidoc for an explanation of test generation.", name);
+            + "See docs/README.asciidoc for an explanation of test generation.", name);
     }
 
     @Override
@@ -96,11 +52,11 @@ public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
 
     @Override
     protected ClientYamlTestClient initClientYamlTestClient(
-            final ClientYamlSuiteRestSpec restSpec,
-            final RestClient restClient,
-            final List<HttpHost> hosts,
-            final Version esVersion,
-            final Version masterVersion) {
+        final ClientYamlSuiteRestSpec restSpec,
+        final RestClient restClient,
+        final List<HttpHost> hosts,
+        final Version esVersion,
+        final Version masterVersion) {
         return new ClientYamlDocsTestClient(restSpec, restClient, hosts, esVersion, masterVersion, this::getClientBuilderWithSniffedHosts);
     }
 
@@ -121,8 +77,9 @@ public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
         return testName != null && (testName.contains("/transform/") || testName.contains("\\transform\\"));
     }
 
-    public void test() throws IOException {
-        runTest(testCandidate);
+    @Override
+    protected CheckedFunction<XContentParser, ExecutableSection, IOException> getParser() {
+        return CompareAnalyzers::parse;
     }
 
     /**
@@ -133,7 +90,7 @@ public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
      * token filters, and should be fairly fast because it compares a fairly
      * small number of tokens.
      */
-    private static class CompareAnalyzers implements ExecutableSection {
+    public static class CompareAnalyzers implements ExecutableSection {
         private static ConstructingObjectParser<CompareAnalyzers, XContentLocation> PARSER =
             new ConstructingObjectParser<>("test_analyzer", false, (a, location) -> {
                 String index = (String) a[0];
@@ -141,15 +98,17 @@ public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
                 String second = (String) a[2];
                 return new CompareAnalyzers(location, index, first, second);
             });
+
         static {
             PARSER.declareString(constructorArg(), new ParseField("index"));
             PARSER.declareString(constructorArg(), new ParseField("first"));
             PARSER.declareString(constructorArg(), new ParseField("second"));
         }
+
         private static CompareAnalyzers parse(XContentParser parser) throws IOException {
             XContentLocation location = parser.getTokenLocation();
             CompareAnalyzers section = PARSER.parse(parser, location);
-            assert parser.currentToken() == Token.END_OBJECT : "End of object required";
+            assert parser.currentToken() == XContentParser.Token.END_OBJECT : "End of object required";
             parser.nextToken(); // throw out the END_OBJECT to conform with other ExecutableSections
             return section;
         }
@@ -170,7 +129,6 @@ public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
         public XContentLocation getLocation() {
             return location;
         }
-
 
 
         @Override
@@ -201,11 +159,11 @@ public class DocsClientYamlTestSuiteIT extends ESClientYamlSuiteTestCase {
             body.put("analyzer", first);
             body.put("text", testText);
             ClientYamlTestResponse response = executionContext.callApi("indices.analyze", singletonMap("index", index),
-                    singletonList(body), emptyMap());
+                singletonList(body), emptyMap());
             Iterator<?> firstTokens = ((List<?>) response.evaluate("tokens")).iterator();
             body.put("analyzer", second);
             response = executionContext.callApi("indices.analyze", singletonMap("index", index),
-                    singletonList(body), emptyMap());
+                singletonList(body), emptyMap());
             Iterator<?> secondTokens = ((List<?>) response.evaluate("tokens")).iterator();
 
             Object previousFirst = null;
